microservices/social-service/requirements.txt
Django>=4.2
psycopg2-binary
requests
python-decouple
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-extensions
celery
requests-oauthlib
python-dotenv 
django-cryptography 
google-api-python-client
google-auth-oauthlib
google-auth 